ThreadRunner: always create new threads
As noted in https://github.com/google/benchmark/issues/461,
Address Space Layout Randomization would be different per-thread,
and that naturally may affect the measurements.

But if we always measure single-thread benchmarks from
the exact same thread, we won't get different ASLR
unless the whole benchmark executable is rerun,
which seems suboptimal.

I'm not 100% sure why we weren't spawning a new thread
in the single-thread case, i mean, is it just
a perf optimization, or are there deeper reasons for it?

Closes https://github.com/google/benchmark/issues/461