python: Assert that libdir is a Path in all branches
The `inplace=True` branch previously relied on some non-trivial interactions
between `str` and `pathlib.Path` objects, which produces unexpected values in
cases of `s / p`, where `s` is a string and `p` is an absolute path.

To fix, just create a `pathlib.Path` of the return value of `build_py.get_package_dir()`.