Guard adding intel compiler flag behind check for intel compiler.
For reasons I don't yet fully understand, failing to guard the check
for flag -wd654 will cause it to be added to certain clang
configurations, and then cause compilation to fail.

While exactly why this happens needs investigation, the fix
proposed in this patch seems direct and simple enough to
apply regardless.