Adding explicit cast to aid in template argument deduction. (#2042)
* Adding explicit cast to aid in template argument deduction.

The `BENCHMARK(my_func)` macro needs to register the function my_func
so the library can run it. It uses `std::make_unique` to create an
object that stores metadata about the benchmark, including its name
("my_func") and a pointer to the function itself (my_func).

The specific constructor being called via `std::make_unique` is for
`benchmark::internal::FunctionBenchmark`.

When a benchmark function name is overloaded, the compiler can't
determine which function is being referred to until the call is made
with specific arguments, or enough context is provided.

The addition of a `static_cast` to the expected `Function*` type
resolves any ambiguity as the compiler is forced to pick the one with
the matching signature.

* clang-format

* add a test

* clang-format again