Make the build work with -Wformat=2 (#2088)
* Make the build work with -Wformat=2

The -Wformat=2 macro requires every format string either be a string
literal, or a parameter that was tagged with the format attribute. The
upshot is that it expects every printf function to be marked, or you get
warnings like this:

    src/console_reporter.cc:104:23: error: format string is not a string literal [-Werror,-Wformat-nonliteral]
      104 |   out << FormatString(fmt, args);
          |                       ^~~

Marking such things is generally worthwhile since it turns on
error-checking within the library, so fill in the missing ones.

Tested with:

    bazelisk build --copt=-Werror --copt=-Wformat=2 :all

* Add -Wformat=2 to catch regressions